// src/interfaces/User.tsx
export interface User {
  mail: string;
  password: string;
  firstname: string;
  lastname: string;
  /*birthdate?: Date;
  gender?: enum;
  path_picture?: string;
  description?: string;
  budget?: enum;
  travel_types?: set;
  lodgings?: set;
  interests?: set;
  available_from?: Date;
  available_to?: Date;
  spoken_languages?: set;
  trip_durations?: set;*/
}